Read ro.hardware.egl for ANGLE's filename
The ANGLE shared object filename is currently hardcoded to
libGLESv2_angle.so, which prevents OEMs from specifying their own
filename when using ANGLE as the default OpenGL ES driver.

This CL updates initializeAnglePlatform() to build the ANGLE library
filename using the suffix specified by ro.hardware.egl when loading
ANGLE as the default OpenGL ES driver.

The filename when loading ANGLE from an APK will remain
libGLESv2_angle.so, for compatibilty reasons. This enforces naming
conventions when loading ANGLE APKs on to the device, regardless of the
name of the built-in version of ANGLE.

Bug: 178871212
Test: Build and launch CF
Merged-In: I462e076fc500d84fa2a27abfa491f82db4a9df80
Change-Id: I59805445fd101707bda867d29c23a25e24f173d8
Signed-off-by: Akash Srivatava <akashniki@gmail.com>
Signed-off-by: Albert I <kras@raphielgang.org>